changed integration tests so both versions are test in the same functions
